backend/requirements.txt
asgiref==3.11.1
colorama==0.4.6
Django==5.2.13
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-django==4.12.0
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.1
